requirements.txt
pytest==8.4.2
requests
jsonschema
allure-pytest==2.15.0
selene==2.0.0rc9
python-dotenv~=1.1.1
selenium==4.38.0